400_splitscreen_scissor
This patch adds a 'scissor' to each of the split screen panes, preventing
them overwriting/overlapping on the otherpanes.

It seems that normally this is not an issue, but is showing with the
Nouveau community drivers and there are lots of comments online that
glViewport does not crop.

Patch is nice and simple...

git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@3703 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: 3a531abf769acc60ff05e828b5c3d8d942d3ba22
Former-commit-id: ff5389d4f3708f5e40af765a8053574777202c5e