281v2_atob
Adds the ability to use a joystick axis for a button type command
(for example ABS or TCS). It automatically selects between button/axis
controls in the case of pressure sensitive buttons.

Uses the ATOB prefix to signify the axis and calibration sets min/max
values. When axis is between these values command is triggered.

Assumes that each axis has a 'null' position, the 'preferences.xml'
will have to be edited if this is not the case.

git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@3710 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: f523e7d497c0ab1bc6da35f61655141a27d0d62d
Former-commit-id: a1de88548b714aa0c562ea73ed470857077e23c8