415_wheel_miscalibration
Abused the pow variable to indicate the polarity of the turn direction
so that the deadzone can be correctly applied.

The joystick stuff now passes 0-1 to either left/right turn, which is
the scaled according to the Pow and Speed sensitivity. There the scaling
is the same regardless of configure/range/deadzone.

Might be worth doing the same for the mouse.

git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@3742 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: 2b5844a16c2160852f53ccdba4e6a58f25c69747
Former-commit-id: 05fff01eaa715c0e187e6f45e191f65de66e6181