Calculate track segment angles at load time. Re #372.
On loading a track, we calculate the straight segments angle's
sine and cosine, so we don't have to do that each time in the RtTrack*
functions.
The changes are introduced in track3.cpp and track4.cpp only, as
older versions are not used in SD.
tTrackSeg struct was added 2 new members, called sin and cos.

git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@3761 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: 252a40dcb7aa2459c17c66e525a5ad25409316c1
Former-commit-id: 9a169e43c5b6ed27d99e427bc1863ed461f5b44c