auto_reverse_during_spin
When auto-reverse is selected reverse can/will be selected if the car spins
this makes recovery difficult.

This patch fixes it so that the 2D speed is used to gate reverse, since the
sqrt(x^2+y^2) is used multiple times it also pre-computes this and stores as
an addition variable. This can/should be used by other functions.

git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@3804 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: beb7b643361d78143e7d3f1b9891bc15dc4015a1
Former-commit-id: 0110389e97e4550dba6e509e26bd19263753f641