track4.cpp: Fix out-of-bounds access on ValStyle
ValStyle is meant to hold a textual representation of every value
defined by enum tSegStyle. However, it is mandatory that both
definitions match because, otherwise, the application could look up
the wrong data or just crash due to a segmentation fault.

So this was what actually happened: ValStyle defined less strings than
enum tSegStyle contained, so an out-of-bounds access was indeed
possible.

In order to avoid this, a X macro, namely TRACK_SEG_STYLES, has been
defined. This allows the definitions for enum tSegStyle and ValStyle
to always match, as they are handled automatically by the preprocessor.

Additionally, it was detected TRK_VAL_PITBUILDING was also missing, and
some elements for ValStyle were wrongly ordered:

- Element [4] should match TRK_VAL_PITBUILDING, not TRK_VAL_FENCE.
- Element [5] should match TRK_VAL_NO_BARRIER, not TRK_VAL_FENCE1.


git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@9409 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: 0888cd18c5cadeae112b80e5c6e388207e4b7c20
Former-commit-id: 9b5a57936367fd898d509a1193aed5430ba2d8cf