Depend on _WIN32 instead of WIN32
Whereas the former is predefined by the compiler [1], the latter is only
defined when including windows.h [2]. Because of the dependency against
windows.h, this had unexpected results where the WIN32 macro was
evaluated before windows.h were included. For example, it caused DLLEXT
to expand to ".so" instead of ".dll", causing unexpected behaviour on
several configuration menus.

In general, the use of the WIN32 macro is considered incorrect [2] and
should be replaced with _WIN32.

[1]: https://learn.microsoft.com/en-us/cpp/preprocessor/predefined-macros?view=msvc-170
[2]: https://stackoverflow.com/questions/662084/whats-the-difference-between-the-win32-and-win32-defines-in-c


git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@9607 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: 7ef940a70e854369cd0669c023347ec18c206998
Former-commit-id: e9ae18d818891f977f49b34bdc59e1fd16e54b3f