Remove robot clones
Robot clones were shared library duplicates with a different name that
the game would rely on to perform different actions based on the car
category.

Not only this was a hacky solution, but it also involved around 50
extra libraries loaded into the game, which means higher resource
consumption and slower startup time, as well as making driver management
more difficult to the much higher of directories involved in
data/drivers/.

Now, only 1 shared library per robot is defined.


Former-commit-id: 4c5141000bc9b46363a16c0928e90b14f40a430a
Former-commit-id: 773f3d6987da3f6ef8e78e0f0276e2d4bf2b07f2