tfg: Add GFPARM_RMODE_NOREREAD
Even if GFPARM_RMODE_REREAD was meant as an optional flag for
GfParmReadFile, it is always set by the implementation. This forced
the parameter file to be always opened for reading, which would fail if
the file did not exist. This is undesired when creating new parameter
files from scratch.

Since this behaviour is probably assumed by many parts of the code, a
new flag, namely GFPARM_RMODE_NOREREAD, has been added so as to disable
such behaviour while maintaining backwards compatibility.


git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@9621 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: 24339c76e7e53a70400f4e24086d5a5f2550a8d8
Former-commit-id: 5bf59aae2e4ddc4d17f95360d1e521d6d5580eda