src/drivers: Remove static limitations
So far, every robot would set up an arbitrary limit on the number of
drivers that could be defined. Not only it was an inefficient approach
because of the fixed-size statically allocated buffers this would incur,
but also there is no reason why this should not be allocated
dynamically.

Now, the internal data structures for each robot are generated
dynamically depending on the number of drivers defined by its parameter
file.


Former-commit-id: 65682fb3d457e2a38fe62cd5b2f00e908bb6aca8
Former-commit-id: 10d86064d066c96f0061080b9ec817e39548277f