shadow: Fix out-of-bounds access
Even if shadow was defined with OOP in mind, for some unknown reason
Driver::Drive relied on a statically allocated array with an arbitrary
size of MAX_MOD_ITF elements (20). Because now the game supports a
virtually unlimited number of drivers (aka interfaces), it was possible
to read past the last_srslip array and cause undefined behaviour.

As hinted above, this array did not make any sense since last_srslip
should be part of each Driver object, and not a statically allocated
array.


Former-commit-id: e365060392b1d08c40a1e3d27f0e113d5a31594d
Former-commit-id: d8954641d78f83d198493e7cb36ffe3db9ef0a55