CMakeLists.txt: Introduce SD_ASSUME_DATADIR
This repository depends on the speed-dreams-data package in order to
generate a working package. However, in the context of CI/CD, it is
desirable to build this package without depending on the
speed-dreams-data, and instead assume it would be found on a standard
location.

Therefore, this would allow the generation of artifacts that install to
standard locations, such as Debian packages.