Fix championship and single event modes
The qualification mode relies on running the race loop as fast as
possible, in order to reproduce the race results in a much shorter time.

However, the maximum frame rate limitations introduced by the commit
below caused a regression to the qualification mode, because now the
game would GfSleep most of the time and only run the race loop once
every frame (e.g.: every 1/60 seconds):

commit 9914e9d24c9eeec75c6f35dbdb3bbb5476d0cb42
Author: xavi92 <xavi92@30fe4595-0a0c-4342-8851-515496e4dcbd>
Date:   Mon Jul 15 02:52:31 2024 +0000

    Implement maximum frame rate limit

Therefore, since it is required to run the physics loop as many times as
possible, the recompute callback variant with `unsigned ms` is used so
that it is run in a loop until the time available to the callback
expires.

Fixes #37 [1].

[1]: https://forge.a-lec.org/speed-dreams/speed-dreams-code/issues/37