3rdParty-devel: Use plib fork
plib is the only library that required a SVN client to fetch the
repository, which is not a sensible requirement any more, now that
the project has been migrated to Git.

Since the snapshots generated by SourceForge are temporary and no
release that matches r2173 is available, we have migrated plib from SVN
to Git and apply our patches to the fork. [1] Therefore, there is no
longer need to keep track of the patches inside the project source tree.

[1]: https://forge.a-lec.org/speed-dreams/plib