easymesh.cpp: Fix non-determinism on call to sqrt(3)
Even if the condition above ensures rho_M is larger or equal than p,
rho_M * rho_M - p * p can still yield two different results when both
values are nearly identical: +0 or -0. Because of how floating-point
arithmetic works, it is not possible to ensure which value would be
generated in a portable way. [1]

If -0 is returned, then sqrt(3) would return nan because of -0 being a
negative number. This invalid calculation propagates to many other
parameters within the source file, causing a cascade of invalid
calculations that end up (typically) with a segmentation fault due to a
wrong memory access.

Since there is no silver bullet to this issue, the use of std::fabs
should at least avoid -0 from being passed to sqrt(3).

[1]: https://bitbashing.io/comparing-floats.html