Fix USR and Shadow not switching beyond 1st gear
When PRM_REALGEARCHANGE is present on the car configuration, this
allows simuv4 and later to simulate more realistic gear changes. This is
achieved by simulating a change to neutral for a specific amount of
time, and finally changing to the target gear as given by
car->ctrl->gear.

However, some driver implementations did not expect such changes to
neutral, and would force a change to the first gear, therefore causing
the driver to loop between both gears.

In order to make drivers such as USR and Shadow aware of this situation,
a new flag, namely car->ctrl->rgcApplying, has been introduced. This
flag tells when such change to neutral is in place, so that drivers do
not change their target gear as long as this flag is set.